Exchangeable Autoregressive Prior

An exchangeable autoregressive prior is used with interactions. An autoregressive model of order \(k\) is applied to the 'along' variable, within each combination of values of the 'by' variables. The damping coefficients are shared across different combinations of the 'by' variables. The series within each combination of the 'by' variables are treated as exchangeable.

Usage

EAR(n = 2, s = 1, along = NULL)

Arguments

n

The order of the model.

s

Scale of half-normal prior for standard deviation (\(\tau\)). Defaults to 1.

along

Name of one of the dimensions in the interaction. Optional, provided the data contain a time or age dimension.

Value

An object of class "bage_prior_ear".

'Along' and 'by' variables

Some priors for interactions distinguish between 'along' and 'by' variables. The 'along' variable is typically time, or, in interactions not involving time, is typically age. The 'by' variables are everything else. In an interaction between region, sex, and time, for instance, the 'along' variable is likely to be time, and the 'by' variables are likely to be region and sex.

If no along argument is supplied, a default value is chosen as follows:

if the interaction contains a time variable, then the 'along' variable is assumed to be time;

otherwise, if the interaction contains an age variable, then the 'along' variable is assumed to be age;

otherwise, an error is raised.

Mathematical details

If \(\beta_{u,v}\) is the \(v\)th element of the interaction within the \(u\)th combination of the 'by' variables, then

$$x_{u,v} = \phi_1 x_{u,v-1} + \cdots + \phi_k x_{u,v-k} + \epsilon_{u,v}$$ $$\epsilon_{u,v} \sim \text{N}(0, \omega^2)$$

where \(\omega\) is chosen so that each \(x_{u,v}\) has marginal variance \(\tau^2\). The value of \(\tau\) has prior

$$\tau \sim \text{N}^+(0, \text{s}^2)$$

The \(\phi_1, \cdots, \phi_k\) are restricted to values within the interval \((-1, 1)\) that yield stationary series.

References

TMB documentation for ARk

See also

Examples

EAR(n = 3)
#> $i_prior
#> [1] 12
#> 
#> $const
#> shape1 shape2    min    max  scale 
#>      2      2     -1      1      1 
#> 
#> $specific
#> $specific$n
#> [1] 3
#> 
#> $specific$shape1
#> [1] 2
#> 
#> $specific$shape2
#> [1] 2
#> 
#> $specific$min
#> [1] -1
#> 
#> $specific$max
#> [1] 1
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "EAR"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_ear" "bage_prior"    
EAR(n = 3, s = 2.4)
#> $i_prior
#> [1] 12
#> 
#> $const
#> shape1 shape2    min    max  scale 
#>    2.0    2.0   -1.0    1.0    2.4 
#> 
#> $specific
#> $specific$n
#> [1] 3
#> 
#> $specific$shape1
#> [1] 2
#> 
#> $specific$shape2
#> [1] 2
#> 
#> $specific$min
#> [1] -1
#> 
#> $specific$max
#> [1] 1
#> 
#> $specific$scale
#> [1] 2.4
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "EAR"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_ear" "bage_prior"    
EAR(along = "cohort")
#> $i_prior
#> [1] 12
#> 
#> $const
#> shape1 shape2    min    max  scale 
#>      2      2     -1      1      1 
#> 
#> $specific
#> $specific$n
#> [1] 2
#> 
#> $specific$shape1
#> [1] 2
#> 
#> $specific$shape2
#> [1] 2
#> 
#> $specific$min
#> [1] -1
#> 
#> $specific$max
#> [1] 1
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> [1] "cohort"
#> 
#> $specific$nm
#> [1] "EAR"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_ear" "bage_prior"