Linear Prior with Autoregressive Errors

Prior for main effect or interaction. A straight line or lines, combined with autoregressive errors. Typically used with time.

Usage

Lin_AR(n_coef = 2, s = 1, sd = 1, along = NULL)

Arguments

n_coef

The number of lagged terms in the model, ie the order of the model. Default is 2.

s

Scale for the innovations in the AR process. Default is 1.

sd

Standard deviation in the prior for the slope of the line. Larger values imply steeper slopes. Default is 1.

along

Name of the variable to be used as the "along" variable. Only used with interactions.

Value

An object of class "bage_prior_linar".

Details

If Lin_AR() is used with an interaction, then, within each combination of the "by" variables, the "along" variable is modelled as a straight line with autoregressive errors.

Argument s controls the size of the errors.

Mathematical details

When Lin_AR() is used with a main effect,

$$\beta_j = \eta q_j + \epsilon_j$$ $$\epsilon_j = \phi_1 \epsilon_{j-1} + \cdots + \phi_n \epsilon_{j-n} + \varepsilon_j,$$ $$\varepsilon_j \sim \text{N}(0, \omega^2).$$

and when it is used with an interaction,

$$\beta_{u,v} = \eta_j q_{u,v} + \epsilon_{u,v}$$ $$\epsilon_{u,v} = \phi_1 \epsilon_{u,v-1} + \cdots + \phi_n \epsilon_{u,v-n} + \varepsilon_{u,v},$$ $$\varepsilon_{u,v} \sim \text{N}(0, \omega^2).$$

where

\(\pmb{\beta}\) is the main effect or interaction;

\(j\) denotes position within the main effect;

\(u\) denotes position within the "along" variable of the interaction;

\(u\) denotes position within the "by" variable(s) of the interaction;

\(n\) is n_coef;

\(q = - (J+1)/(J-1) + 2j/(J-1);\) and

\(q_v = - (V+1)/(V-1) + 2v/(V-1)\).

The slopes have priors $$\eta \sim \text{N}(0, \text{sd}^2)$$ and $$\eta_u \sim \text{N}(0, \text{sd}^2).$$ Larger values for sd permit steeper slopes.

Internally, Lin_AR() derives a value for \(\omega\) that gives \(\epsilon_j\) or \(\epsilon_{u,v}\) a marginal variance of \(\tau^2\). Parameter \(\tau\) has a half-normal prior $$\tau \sim \text{N}^+(0, \text{s}^2),$$ where a value for s is provided by the user.

The \(\phi_1, \cdots, \phi_k\) are restricted to values between -1 and 1 that jointly lead to a stationary model. The quantity \(r = \sqrt{\phi_1^2 + \cdots + \phi_k^2}\) is given a boundary-avoiding prior

$$r \sim \text{Beta}(2, 2).$$

Slope \(\eta\) is drawn from a normal distribution, $$\eta \sim \text{N}(0, (\text{sd})^2).$$

See also

Examples

Lin_AR()
#>   Lin_AR() 
#>   n_coef: 2
#>        s: 1
#>       sd: 1
#>      min: -1
#>      max: 1
#>    along: NULL
Lin_AR(n_coef = 3, s = 0.5, sd = 2)
#>   Lin_AR(n_coef=3,s=0.5,sd=2) 
#>   n_coef: 3
#>        s: 0.5
#>       sd: 2
#>      min: -1
#>      max: 1
#>    along: NULL