Random Walk Prior, with Seasonal Effect

Prior for a main effect or interaction, typically involving time. Combines a random walk with a seasonal effect.

Usage

RWSeas(n, s = 1, s_seas = 1, along = NULL)

Arguments

n

Number of seasons

s

Scale for prior for innovations in the random walk. Default is 1.

s_seas

Scale for prior for innovations in the seasonal effect. Default is 1. Can be 0.

along

Name of the variable to be used as the "along" variable. Only used with interactions.

Value

Object of class "bage_prior_rwseasvary"

or "bage_prior_rwseasfix".

Details

If RWSeas() is used with an interaction, then separate series are used for the "along" variable within each combination of the "by" variables.

Argument s controls the size of innovations in the random walk. Smaller values for s tend to give smoother series.

Argument n controls the number of seasons. When using quarterly data, for instance, n should be 4, and when using monthly data, n should be 12.

By default, the magnitude of seasonal effects can change over time. However, setting s_seas to 0 produces seasonal effects that are "fixed", ie that are the same every year.

Mathematical details

When RWSeas() is used with a main effect,

$$\beta_j = \alpha_j + \lambda_j$$ $$\alpha_j \sim \text{N}(\alpha_{j-1}, \tau^2)$$ $$\lambda_j \sim \text{N}(\lambda_{j-n}, \omega^2),$$

and when it is used with an interaction,

$$\beta_{u,v} = \alpha_{u,v} + \lambda_{u,v}$$ $$\alpha_{u,v} \sim \text{N}(\alpha_{u,v-1}, \tau^2),$$ $$\lambda_{u,v} \sim \text{N}(\lambda_{u,v-n}, \omega^2)$$

where

\(\pmb{\beta}\) is the main effect or interaction;

\(\alpha_j\) or \(\alpha_{u,v}\) is an element of the random walk;

\(\lambda_j\) or \(\lambda_{u,v}\) is an element of the seasonal effect;

\(j\) denotes position within the main effect;

\(v\) denotes position within the "along" variable of the interaction;

\(u\) denotes position within the "by" variable(s) of the interaction; and

\(n\) is the number of seasons.

Parameter \(\omega\) has a half-normal prior $$\omega \sim \text{N}^+(0, \text{s\_seas}^2),$$ where s_seas is provided by the user. If s_seas is set to 0, then \(\omega\) is 0, and the seasonal effects are fixed over time.

Parameter \(\tau\) has a half-normal prior $$\tau \sim \text{N}^+(0, \text{s}^2),$$ where s is provided by the user.

See also

Examples

RWSeas(n = 4)             ## seasonal effects evolve
#> $i_prior
#> [1] 11
#> 
#> $const
#>          n scale_seas      scale 
#>          4          1          1 
#> 
#> $specific
#> $specific$n
#> [1] 4
#> 
#> $specific$scale_seas
#> [1] 1
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_rwseasvary" "bage_prior"           
RWSeas(n = 4, s_seas = 0) ## seasonal effects fixed
#> $i_prior
#> [1] 10
#> 
#> $const
#>     n scale 
#>     4     1 
#> 
#> $specific
#> $specific$n
#> [1] 4
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_rwseasfix" "bage_prior"