SVD Prior for Age

A Singular Value Decomposition (SVD) prior for a main effect or interaction involving age.

Usage

SVD(ssvd, n = NULL)

Arguments

ssvd

Object of class "bage_ssvd" holding a scaled SVD. See below for scaled SVDs of databases currently available in bage.

n

Number of components from scaled SVD to use in modelling. The default is half the number of components of ssvd.

Value

An object of class "bage_prior_svd".

Details

A SVD() prior assumes that the age profile for the quantity being modelled looks like it was drawn at random from an external demographic database. For instance, the SVD() prior obtained via

SVD(HMD)

favours age profiles that look like they were obtained from the Human Mortality Database.

If SVD() is used with an interaction, then separate profiles are constructed along within each combination of the "by" variables.

Age-sex or age-gender profiles should be modelled using SVDS() rather than SVD().

Mathematical details

When SVD() is used with a main effect,

$$\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g},$$

and when it is used with an interaction,

$$\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g},$$

where

\(\pmb{\beta}\) is a main effect or interaction with \(J\) elements;

\(\pmb{\beta}_u\) is a subvector of \(\pmb{\beta}\) holding values for the \(u\)th combination of the "by" variables;

\(J\) is the number of elements of \(\pmb{\beta}\);

\(V\) is the number of elements of \(\pmb{\beta}_u\);

\(\pmb{F}\) is a known matrix with dimension \(J \times n\) if \(\pmb{\beta}\) is a main effect and dimension \(V \times n\) if \(\pmb{\beta}\) is an interaction; and

\(\pmb{g}\) is a known vector with \(J\) elements if if \(\pmb{\beta}\) is a main effect and \(V\) elements if \(\pmb{\beta}\) is an interaction.

\(\pmb{F}\) and \(\pmb{g}\) are constructed from a large database of age-specific demographic estimates by performing an SVD and standardizing.

The elements of \(\pmb{\alpha}\) have prior $$\alpha_k \sim \text{N}(0, 1), \quad k = 1, \cdots, K.$$

Scaled SVDs of demographic databases in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

References

For details of the construction of scaled SVDS see the vignette here.

See also

Examples

SVD(HMD)
#> Loading required package: Matrix
#> $i_prior
#> [1] 9
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> integer(0)
#> 
#> $specific$joint
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_svd" "bage_prior"    
SVD(HMD, n = 3)
#> $i_prior
#> [1] 9
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> [1] 3
#> 
#> $specific$joint
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_svd" "bage_prior"