SVD Prior for Age-Sex or Age-Gender

A Singular Value Decomposition (SVD) prior for an interaction involving age-sex profiles or age-gender profiles.

Usage

SVDS(ssvd, n = 5, joint = FALSE)

Arguments

ssvd

Object of class "bage_ssvd" holding a scaled SVD. See below for scaled SVDs of databases currently available in bage.

n

Number of components from scaled SVD to use in modelling. The default is half the number of components of ssvd.

joint

Whether to use combined or separate SVDs. Default is FALSE. See below for details.

Value

An object of class "bage_prior_svd".

Details

A SVDS() prior assumes that the age-sex profile or age-gender profile for the quantity being modelled looks like it was drawn at random from an external demographic database. For instance, the prior obtained via

SVDS(HMD)

favours age-sex profiles that look like they were obtained from the Human Mortality Database.

If SVDS() is used with an interaction between age, sex or gender, and one or more "by" variables, then separate age-sex or age-gender profiles are constructed within each combination of the "by" variables.

Age profiles with no sex or gender dimension should be modelled using SVD() rather than SVDS().

Joint or independent SVDs

Two possible ways of extracting patterns from age-sex-specific data are

carry out separate SVDs on separate datasets for each sex/gender; or

carry out a single SVD on dataset that has separate entries for each sex/gender.

Option 1 is more flexible. Option 2 is more robust to sampling or measurement errors. Option 1 is obtained by setting the joint argument to FALSE. Option 2 is obtained by setting the joint argument to TRUE. The default is FALSE.

Mathematical details

Case 1: Interaction involving age and sex or gender, but no other variables

When SVDS() is used with joint = FALSE,

$$\pmb{\beta}_{s} = \pmb{F}_s \pmb{\alpha}_{s} + \pmb{g}_s,$$

and when SVDS() is used with joint = TRUE,

$$\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction;

\(\pmb{\beta}_{s}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\), with \(s = 1, \dots, S\);

\(J\) is the number of elements in \(\pmb{\beta}\);

\(\pmb{F}_s\) is a known \((J/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(J/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(J \times n\) matrix, with values for all sexes/genders.

\(\pmb{g}\) is a known vector with \(J\) elements, with values for all sexes/genders.

The elements of \(\pmb{\alpha}_s\) and \(\pmb{\alpha}\) have prior $$\alpha_k \sim \text{N}(0, 1).$$

Case 2: Interaction involving age and sex or gender, and one or more other variables

When SVDS() is used with joint = FALSE,

$$\pmb{\beta}_{u,s} = \pmb{F}_s \pmb{\alpha}_{u,s} + \pmb{g}_s,$$

and when SVDS() is used with joint = TRUE,

$$\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction;

\(\pmb{\beta}_{u,s}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\) for the \(u\)th combination of "by" variables, with \(s = 1, \dots, S\);

\(V\) is the number of elements in \(\pmb{\beta}_u\);

\(\pmb{F}_s\) is a known \((V/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(V/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(V \times n\) matrix, with values for all sexes/genders.

\(\pmb{g}\) is a known vector with \(V\) elements, with values for all sexes/genders.

The elements of \(\pmb{\alpha}_{u,s}\) and \(\pmb{\alpha}_u\) have prior $$\alpha_k \sim \text{N}(0, 1).$$

Scaled SVDs of demographic databases in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

References

For details of the construction of scaled SVDS see the vignette here.

See also

Examples

SVDS(HMD)
#> $i_prior
#> [1] 9
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> [1] 5
#> 
#> $specific$joint
#> [1] FALSE
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_svd" "bage_prior"    
SVDS(HMD, joint = TRUE)
#> $i_prior
#> [1] 9
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> [1] 5
#> 
#> $specific$joint
#> [1] TRUE
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_svd" "bage_prior"