SVD Priors for Age-Sex or Age-Gender Interactions, With Time-Varying Coefficients

Singular Value Decomposition (SVD) priors for interactions involving age, sex or gender, time, and possibly other variables.

Usage

SVDS_AR(ssvd, n_comp = NULL, joint = FALSE, n_coef = 2, s = 1)

SVDS_AR1(ssvd, n_comp = NULL, joint = FALSE, min = 0.8, max = 0.98, s = 1)

SVDS_RW(ssvd, n_comp = NULL, joint = FALSE, s = 1)

SVDS_RW2(ssvd, n_comp = NULL, joint = FALSE, s = 1)

Arguments

ssvd

Object of class "bage_ssvd" holding a scaled SVD. See below for scaled SVDs of databases currently available in bage.

n_comp

Number of components from scaled SVD to use in modelling. The default is half the number of components of ssvd.

joint

Whether to use combined or separate SVDs. Default is FALSE. See below for details.

n_coef

Number of AR coefficients in SVD_RW().

s

Scale for standard deviations terms.

min, max

Minimum and maximum values for autocorrelation coefficient in SVD_AR(). Defaults are 0.8 and 0.98.

Value

An object of class "bage_prior_svd_ar", "bage_prior_svd_rw", or "bage_prior_svd_rw2".

Details

SVD_AR(), SVD_AR1(), SVD_RW(), and SVD_RW2() priors assume that, in any given period, the age-sex profile or age-gender profiles for the quantity being modelled look they were drawn at random from an external demographic database. For instance, the prior obtained via

SVDS_AR1(HMD)

favours age-sex profiles that look like they were obtained from the Human Mortality Database.

Time-varying age profiles with no sex or gender dimension should be modelled using SVD_AR(), SVD_AR1(), SVD_RW(), or SVD_RW2() priors, rather than SVDS_AR(), SVDS_AR1(), SVDS_RW(), or SVDS_RW2().

Mathematical details

Case 1: Interaction involving age, sex/gender, time, and no other variables

When joint is FALSE,

$$\pmb{\beta}_{s,t} = \pmb{F}_s \pmb{\alpha}_{s,t} + \pmb{g}_s,$$

and when joint is TRUE,

$$\pmb{\beta}_t = \pmb{F} \pmb{\alpha}_t + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction involving age, sex/gender, and time;

\(\pmb{\beta}_{s,t}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\) in period \(t\);

\(\pmb{\beta}_t\) is a subvector of \(\pmb{\beta}\), holding values for all combinations of age and sex/gender in period \(t\);

\(J\) is the number of elements in \(\pmb{\beta}_t\);

\(S\) is the number of sexes/genders;

\(n\) is n_comp;

\(\pmb{F}_s\) is a known \((J/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(J/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(J \times n\) matrix, with values for all combinations of age and sex/gender; and

\(\pmb{g}\) is a known vector with \(J\) elements, with values for all combinations of age and sex/gender.

Case 2: Interaction involving age, sex/gender, time, and one or more other variables

When joint is FALSE,

$$\pmb{\beta}_{u,s,t} = \pmb{F}_s \pmb{\alpha}_{u,s,t} + \pmb{g}_s,$$

and when joint is TRUE,

$$\pmb{\beta}_{u,t} = \pmb{F} \pmb{\alpha}_{u,t} + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction involving sex/gender;

\(\pmb{\beta}_{u,s,t}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\), for the \(u\)th combination of the other variables, in period \(t\);

\(\pmb{\beta}_{u,t}\) is a subvector of \(\pmb{\beta}\), holding values for all combinatinos of age and sex/gender, for the \(u\)th combination of the other variables, in period \(t\);

\(V\) is the number of elements in \(\pmb{\beta}_{u,t}\);

\(S\) is the number of sexes/genders;

\(n\) is n_comp;

\(\pmb{F}_s\) is a known \((V/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(V/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(V \times n\) matrix, with values for all combinations of age and sex/gender; and

\(\pmb{g}\) is a known vector with \(V\) elements, with values for all commbinations of age and sex/gender.

Priors for coefficients

With SVDS_AR(), the prior for the \(k\)th elements of \(\pmb{\alpha}_{s,t}\) or \(\pmb{\alpha}_{u,s,t}\) is

$$\alpha_{k,s,t} = \phi_1 \alpha_{k,s,t-1} + \cdots + \phi_n \beta_{k,s,t-n} + \epsilon_{k,s,t}$$

or

$$\alpha_{k,u,s,t} = \phi_1 \alpha_{k,u,s,t-1} + \cdots + \phi_n \beta_{k,u,s,t-n} + \epsilon_{k,u,t}$$;

with SVDS_AR1(), it is

$$\alpha_{k,s,t} = \phi \alpha_{k,s,t-1} + \epsilon_{k,s,t}$$

or

$$\alpha_{k,u,s,t} = \phi \alpha_{k,u,s,t-1} + \epsilon_{k,u,s,t}$$;

with SVDS_RW(), it is

$$\alpha_{k,s,t} = \alpha_{k,s,t-1} + \epsilon_{k,s,t}$$

or

$$\alpha_{k,u,s,t} = \alpha_{k,u,s,t-1} + \epsilon_{k,u,s,t}$$;

and with SVDS_RW2(), it is

$$\alpha_{k,s,t} = 2 \alpha_{k,s,t-1} - \alpha_{k,s,t-2} + \epsilon_{k,s,t}$$

or

$$\alpha_{k,u,s,t} = 2 \alpha_{k,u,s,t-1} - \alpha_{k,u,s,t-2} + \epsilon_{k,u,s,t}$$.

For more on the \(\phi\) and \(\epsilon\), see AR(), AR1(), RW(), and RW2().

Joint or independent SVDs

Two possible ways of extracting patterns from age-sex-specific data are

carry out separate SVDs on separate datasets for each sex/gender; or

carry out a single SVD on dataset that has separate entries for each sex/gender.

Option 1 is more flexible. Option 2 is more robust to sampling or measurement errors. Option 1 is obtained by setting the joint argument to FALSE. Option 2 is obtained by setting the joint argument to TRUE. The default is FALSE.

Scaled SVDs of demographic databases in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

References

For details of the construction of scaled SVDS see the vignette here.

See also

Examples

SVDS_AR1(HMD)
#>   SVDS_AR1(HMD) 
#>     ssvd: HMD
#>   n_comp: 5
#>    joint: FALSE
#>      min: 0.8
#>      max: 0.98
#>        s: 1
#>    along: NULL
SVDS_RW(HMD, n_comp = 3)
#>   SVDS_RW(HMD,n_comp=3) 
#>     ssvd: HMD
#>   n_comp: 3
#>    joint: FALSE
#>        s: 1