P-Spline Prior

A p-spline (penalised spine) prior for main effects or interactions. Typically used with age, but can be used with any variable where outcomes are expected to vary smoothly from one element to the next.

Usage

Sp(n = NULL, s = 1, along = NULL)

Arguments

n

Number of spline basis functions to use.

s

Scale for the prior for the innovations. Default is 1.

along

Name of the variable to be used as the "along" variable. Only used with interactions.

Value

An object of class "bage_prior_spline".

Details

If Sp() is used with an interaction, then separate splines are used for the "along" variable within each combination of the "by" variables.

Mathematical details

When Sp() is used with a main effect,

$$\pmb{\beta} = \pmb{X} \pmb{\alpha}$$

and when it is used with an interaction,

$$\pmb{\beta}_u = \pmb{X} \pmb{\alpha}_u$$

where

\(\pmb{\beta}\) is the main effect or interaction, with \(J\) elements;

\(\pmb{beta}_u\) is a subvector of \(\pmb{\beta}\) holding values for the \(u\)th combination of the "by" variables;

\(J\) is the number of elements of \(\pmb{\beta}\);

\(V\) is the number of elements of \(\pmb{\beta}_u\); and

\(X\) is a \(J \times n\) or \(V \times n\) matrix of spline basis functions.

The elements of \(\pmb{\alpha}\) or \(\pmb{alpha}_u\) are assumed to follow a random walk with drift.

References

Eilers, P.H.C. and Marx B. (1996). "Flexible smoothing with B-splines and penalties". Statistical Science. 11 (2): 89–121.

See also

Examples

Sp()
#> $i_prior
#> [1] 8
#> 
#> $const
#> scale 
#>     1 
#> 
#> $specific
#> $specific$n
#> NULL
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_spline" "bage_prior"       
Sp(n = 10)
#> $i_prior
#> [1] 8
#> 
#> $const
#> scale 
#>     1 
#> 
#> $specific
#> $specific$n
#> [1] 10
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_spline" "bage_prior"