Extract Data and Modelled Values

Extract data and rates, probabilities, or means from a model object. The return value consists of the original data and one or more columns of modelled values.

Usage

# S3 method for class 'bage_mod'
augment(x, quiet = FALSE, ...)

Arguments

x

An object of class "bage_mod".

quiet

Whether to suppress messages. Default is FALSE.

...

Unused. Included for generic consistency only.

Value

A tibble, with the original data plus the following columns:

.observed 'Direct' estimates of rates or probabilities, ie counts divided by exposure or size (in Poisson and binomial models.)

.fitted Draws of rates, probabilities, or means.

.expected Draws of expected values for rates or probabilities (in Poisson that include exposure, or in binomial models.)

Uncertain quantities are represented using rvecs.

Fitted vs unfitted models

augment() is typically called on a fitted model. In this case, the modelled values are draws from the joint posterior distribution for rates, probabilities, or means.

augment() can, however, be called on an unfitted model. In this case, the modelled values are draws from the joint prior distribution. In other words, the modelled values are informed by model priors, and by values for exposure, size, or weights, but not by observed outcomes.

See also

Examples

## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> augment()
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 912 × 9
#>    age   sex    ethnicity  year             injuries  popn           .observed
#>    <fct> <chr>  <chr>     <int>         <rdbl<1000>> <int>        <rdbl<1000>>
#>  1 0-4   Female Maori      2000  22551 (28, 1859660) 35830  0.63 (0.00078, 52)
#>  2 5-9   Female Maori      2000  22724 (34, 2290327) 35120  0.65 (0.00097, 65)
#>  3 10-14 Female Maori      2000  19179 (35, 3753925) 32830  0.58 (0.0011, 114)
#>  4 15-19 Female Maori      2000  16806 (21, 5084151) 27130 0.62 (0.00077, 187)
#>  5 20-24 Female Maori      2000   12640 (7, 6635266) 24380 0.52 (0.00029, 272)
#>  6 25-29 Female Maori      2000     13831 (4, 8e+06) 24160 0.57 (0.00016, 332)
#>  7 30-34 Female Maori      2000   12724 (2, 6851652) 22560 0.56 (8.8e-05, 304)
#>  8 35-39 Female Maori      2000   11911 (1, 1.4e+07) 22230 0.54 (4.5e-05, 613)
#>  9 40-44 Female Maori      2000    9550 (1, 1.2e+07) 18130 0.53 (5.5e-05, 647)
#> 10 45-49 Female Maori      2000 7794 (0.98, 1.1e+07) 13770 0.57 (7.1e-05, 796)
#> # ℹ 902 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> augment()
#> # A tibble: 912 × 9
#>    age   sex    ethnicity  year injuries  popn .observed
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>
#>  1 0-4   Female Maori      2000       12 35830 0.000335 
#>  2 5-9   Female Maori      2000        6 35120 0.000171 
#>  3 10-14 Female Maori      2000        3 32830 0.0000914
#>  4 15-19 Female Maori      2000        6 27130 0.000221 
#>  5 20-24 Female Maori      2000        6 24380 0.000246 
#>  6 25-29 Female Maori      2000        6 24160 0.000248 
#>  7 30-34 Female Maori      2000       12 22560 0.000532 
#>  8 35-39 Female Maori      2000        3 22230 0.000135 
#>  9 40-44 Female Maori      2000        6 18130 0.000331 
#> 10 45-49 Female Maori      2000        6 13770 0.000436 
#> # ℹ 902 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>