Extract Values for Hyper-Parameters

Extract values for hyper-parameters from a model object. Hyper-parameters include main effects and interactions, dispersion and variance terms, and SVD or spline coefficients.

Usage

# S3 method for class 'bage_mod'
components(object, quiet = FALSE, ...)

Arguments

object

Object of class "bage_mod", typically created with mod_pois(), mod_binom(), or mod_norm().

quiet

Whether to suppress messages. Default is FALSE.

...

Unused. Included for generic consistency only.

Value

A tibble with four columns columns:

The return value contains the following columns:

term Model term that the hyper-parameter belongs to.

component Component within term.

level Element within component .

.fitted An rvec containing draws from the posterior distribution.

Fitted vs unfitted models

components() is typically called on a fitted model. In this case, the modelled values are draws from the joint posterior distribution for the hyper-parameters in the model.

components() can, however, be called on an unfitted model. In this case, the modelled values are draws from the joint prior distribution. In other words, the modelled values are informed by model priors, and by any exposure, size, or weights argument in the model, but not by the observed outcomes.

See also

Examples

## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = nzl_injuries,
                exposure = popn)

## extract prior distribution
## of hyper-parameters
mod |>
  components()
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 37 × 4
#>    term        component level                  .fitted
#>    <chr>       <chr>     <chr>             <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)   -0.015 (-2, 1.8)
#>  2 age         effect    0-4           -0.044 (-1.9, 2)
#>  3 age         effect    5-9         -0.045 (-2.8, 2.9)
#>  4 age         effect    10-14       -0.035 (-3.5, 3.9)
#>  5 age         effect    15-19       0.0017 (-4.1, 4.5)
#>  6 age         effect    20-24        0.021 (-4.8, 5.2)
#>  7 age         effect    25-29       0.0062 (-5.2, 5.4)
#>  8 age         effect    30-34        0.058 (-5.6, 6.1)
#>  9 age         effect    35-39        0.046 (-5.8, 6.1)
#> 10 age         effect    40-44        0.031 (-6.5, 6.7)
#> # ℹ 27 more rows

## fit model
mod <- mod |>
  fit()

## extract posterior distribution
## of hyper-parameters
mod |>
  components()
#> # A tibble: 37 × 4
#>    term        component level                   .fitted
#>    <chr>       <chr>     <chr>              <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)  -2.6 (-4.1, -0.95)
#>  2 age         effect    0-4          -2.5 (-4.1, -0.87)
#>  3 age         effect    5-9           -3.8 (-5.4, -2.1)
#>  4 age         effect    10-14         -3.4 (-4.9, -1.8)
#>  5 age         effect    15-19        -1.6 (-3.2, 0.017)
#>  6 age         effect    20-24           -1.5 (-3, 0.17)
#>  7 age         effect    25-29        -1.6 (-3.2, 0.018)
#>  8 age         effect    30-34       -1.7 (-3.3, -0.082)
#>  9 age         effect    35-39       -1.7 (-3.3, -0.059)
#> 10 age         effect    40-44        -1.7 (-3.3, -0.12)
#> # ℹ 27 more rows