Extract Components used by SVD Summary

Extract the matrix and offset used by a scaled SVD summary of a demographic database.

Usage

# S3 method for bage_ssvd
components(object, n_comp = NULL, joint = NULL, age_labels = NULL, ...)

Arguments

object

An object of class "bage_ssvd".

n_comp

The number of components. The default is half the total number of components of object.

joint

Whether to use joint or independent SVDs for each sex/gender. If no value is supplied, an SVD with no sex/gender dimension is used.

age_labels

Age labels for the desired age or age-sex profile. If no labels are supplied, the most detailed profile available is used.

...

Not currently used.

Value

A tibble with the offset and components.

Scaled SVDs in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

See also

Examples

## females and males combined
components(LFP, n_comp = 3)
#> # A tibble: 52 × 3
#>    component age    value
#>    <fct>     <chr>  <dbl>
#>  1 Offset    15-19 -0.789
#>  2 Offset    20-24  0.768
#>  3 Offset    25-29  1.45 
#>  4 Offset    30-34  1.55 
#>  5 Offset    35-39  1.60 
#>  6 Offset    40-44  1.63 
#>  7 Offset    45-49  1.52 
#>  8 Offset    50-54  1.19 
#>  9 Offset    55-59  0.594
#> 10 Offset    60-64 -0.405
#> # ℹ 42 more rows

## joint model for females and males
components(LFP, joint = TRUE, n_comp = 3)
#> # A tibble: 104 × 4
#>    component sex    age     value
#>    <fct>     <chr>  <chr>   <dbl>
#>  1 Offset    Female 15-19 -0.917 
#>  2 Offset    Female 20-24  0.489 
#>  3 Offset    Female 25-29  0.799 
#>  4 Offset    Female 30-34  0.743 
#>  5 Offset    Female 35-39  0.847 
#>  6 Offset    Female 40-44  0.959 
#>  7 Offset    Female 45-49  0.901 
#>  8 Offset    Female 50-54  0.606 
#>  9 Offset    Female 55-59  0.0367
#> 10 Offset    Female 60-64 -0.920 
#> # ℹ 94 more rows

## females and males modelled independently
components(LFP, joint = FALSE, n_comp = 3)
#> # A tibble: 104 × 4
#>    component sex    age     value
#>    <fct>     <chr>  <chr>   <dbl>
#>  1 Offset    Female 15-19 -0.916 
#>  2 Offset    Female 20-24  0.481 
#>  3 Offset    Female 25-29  0.810 
#>  4 Offset    Female 30-34  0.763 
#>  5 Offset    Female 35-39  0.855 
#>  6 Offset    Female 40-44  0.953 
#>  7 Offset    Female 45-49  0.889 
#>  8 Offset    Female 50-54  0.597 
#>  9 Offset    Female 55-59  0.0388
#> 10 Offset    Female 60-64 -0.907 
#> # ℹ 94 more rows

## specify age groups
labels <- poputils::age_labels(type = "five", min = 15, max = 60)
components(LFP, age_labels = labels)
#> # A tibble: 36 × 3
#>    component   age    value
#>    <fct>       <chr>  <dbl>
#>  1 Offset      15-19 -1.00 
#>  2 Offset      20-24  0.738
#>  3 Offset      25-29  1.48 
#>  4 Offset      30-34  1.63 
#>  5 Offset      35-39  1.70 
#>  6 Offset      40-44  1.73 
#>  7 Offset      45-49  1.61 
#>  8 Offset      50-54  1.26 
#>  9 Offset      55-59  0.629
#> 10 Component 1 15-19 -0.281
#> # ℹ 26 more rows