Compose a Prior for a Time Main Effect or Interaction

Create a composite prior for a time main effect, or for an interaction involving time. The composite prior always contains a trend, and may contain a cyclical effect, a seasonal effect, and an error.

Usage

compose_time(trend, cyclical = NULL, seasonal = NULL, error = NULL)

Arguments

trend

Prior describing the long-run behavior of the series. See below for choices. Required.

cyclical

Prior describing departures from the long-run trend that do not have a fixed period. See below for choices. Optional.

seasonal

Prior describing departures from the long-run trend that have a fixed period. See below for choices. Optional.

error

Prior for additional idiosyncratic variation. See below for choices. Optional

Value

An object of class "bage_prior_compose"

Details

Term	Main effects	Interactions

trend	Lin(), RW(), RW2(), Sp()	ELin()

seasonal	Seas()	ESeas()

cyclical	AR(), AR1()	EAR(), EAR1()

error	N()	N()

Examples

compose_time(
  trend = Lin(),
  cyclical = AR()
)
#> $i_prior
#> [1] 1000
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$priors
#> $specific$priors$trend
#> $i_prior
#> [1] 8
#> 
#> $const
#>    scale sd_slope 
#>        1        1 
#> 
#> $specific
#> $specific$scale
#> [1] 1
#> 
#> $specific$sd_slope
#> [1] 1
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_lin" "bage_prior"    
#> 
#> $specific$priors$cyclical
#> $i_prior
#> [1] 5
#> 
#> $const
#> shape1 shape2    min    max  scale 
#>      3      3     -1      1      1 
#> 
#> $specific
#> $specific$n
#> [1] 2
#> 
#> $specific$shape1
#> [1] 3
#> 
#> $specific$shape2
#> [1] 3
#> 
#> $specific$min
#> [1] -1
#> 
#> $specific$max
#> [1] 1
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$nm
#> [1] "AR"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_ar" "bage_prior"   
#> 
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "compose_time"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_compose" "bage_prior"        

compose_time(
  trend = RW2(),
  error = N()
)
#> $i_prior
#> [1] 1000
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$priors
#> $specific$priors$trend
#> $i_prior
#> [1] 4
#> 
#> $const
#>    scale sd_slope 
#>        1        1 
#> 
#> $specific
#> $specific$scale
#> [1] 1
#> 
#> $specific$sd_slope
#> [1] 1
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_rw2" "bage_prior"    
#> 
#> $specific$priors$error
#> $i_prior
#> [1] 1
#> 
#> $const
#> scale 
#>     1 
#> 
#> $specific
#> $specific$scale
#> [1] 1
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_norm" "bage_prior"     
#> 
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "compose_time"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_compose" "bage_prior"        

compose_time(
  trend = ELin(),
  cyclical = EAR(),
  season = ESeas(n = 4)
)
#> $i_prior
#> [1] 1000
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$priors
#> $specific$priors$trend
#> $i_prior
#> [1] 9
#> 
#> $const
#>    scale sd_slope   mscale 
#>        1        1        1 
#> 
#> $specific
#> $specific$scale
#> [1] 1
#> 
#> $specific$sd_slope
#> [1] 1
#> 
#> $specific$mscale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_elin" "bage_prior"     
#> 
#> $specific$priors$cyclical
#> $i_prior
#> [1] 12
#> 
#> $const
#> shape1 shape2    min    max  scale 
#>      3      3     -1      1      1 
#> 
#> $specific
#> $specific$n
#> [1] 2
#> 
#> $specific$shape1
#> [1] 3
#> 
#> $specific$shape2
#> [1] 3
#> 
#> $specific$min
#> [1] -1
#> 
#> $specific$max
#> [1] 1
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "EAR"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_ear" "bage_prior"    
#> 
#> $specific$priors$seasonal
#> $i_prior
#> [1] 11
#> 
#> $const
#>    scale <unused> <unused> <unused> 
#>        1        0        0        0 
#> 
#> $specific
#> $specific$n
#> [1] 4
#> 
#> $specific$scale
#> [1] 1
#> 
#> $specific$along
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_eseas" "bage_prior"      
#> 
#> 
#> $specific$along
#> NULL
#> 
#> $specific$nm
#> [1] "compose_time"
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_compose" "bage_prior"