Use a Model to Make a Forecast

Forecast rates, probabilities, or means, and other model parameters.

Usage

# S3 method for bage_mod
forecast(
  object,
  output = c("augment", "components"),
  include_estimates = FALSE,
  labels,
  ...
)

Arguments

object

An object of class "bage_mod".

output

Type of output returned

include_estimates

Whether to include historical estimates along with the forecasts. Default is FALSE.

labels

Labels for future values. WARNING this argument is only temporary.

...

Not currently used.

Value

A tibble.

How the forecasts are constructed

Internally, the steps involved in a forecast are:

Forecast time-varying main effects and interactions, e.g. a time main effect, or an age-time interaction.

Combine forecasts for the time-varying main effects and interactions with non-time-varying parameters, e.g. age effects or dispersion.

Use the combined parameters to generate values for rates, probabilities or means.

See LINK TO VIGNETTE for more details.

Output

When output is "augment" (the default), the return value from forecast() looks like output from function augment(). When output is "components", the return value looks like output from components().

When include_estimates is FALSE (the default), the output of forecast() excludes values for time-varying parameters for the period covered by the data. When include_estimates is TRUE, the output includes these values. Setting include_estimates to TRUE can be helpful when creating graphs that combine estimates and forecasts.

Using data in forecasts

forecast() is typically used with a fitted model, i.e. a model in which parameter values have been estimated from the data. The resulting forecasts reflect data and priors.

forecast() can, however, be used with an unfitted model. In this case, the forecasts are based entirely on the priors. See below for an example. Experimenting with forecasts based entirely on the priors can be helpful for designing a model. See LINK TO WORKFLOW VIGNETTE for details.

Warning

The interface for forecast() has not been finalised and is likely to change in the near future.

See also

Examples

## specify and fit model
mod <- mod_pois(injuries ~ age * sex + ethnicity + year,
                data = injuries,
                exposure = popn) |>
  fit(mod)
mod
#> -- Fitted Poisson model --
#> 
#>    injuries ~ age * sex + ethnicity + year
#> 
#> (Intercept) ~ NFix()
#>         age ~ RW()
#>         sex ~ NFix()
#>   ethnicity ~ NFix()
#>        year ~ RW()
#>     age:sex ~ RW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>   var_sexgender: sex
#>        var_time: year
#>          n_draw: 1000

## forecasts
mod |>
  forecast(labels = 2019:2024)
#> # A tibble: 288 × 9
#>    age   sex    ethnicity  year injuries  popn .observed
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>
#>  1 0-4   Female Maori      2019       NA    NA        NA
#>  2 0-4   Female Maori      2020       NA    NA        NA
#>  3 0-4   Female Maori      2021       NA    NA        NA
#>  4 0-4   Female Maori      2022       NA    NA        NA
#>  5 0-4   Female Maori      2023       NA    NA        NA
#>  6 0-4   Female Maori      2024       NA    NA        NA
#>  7 0-4   Female Non Maori  2019       NA    NA        NA
#>  8 0-4   Female Non Maori  2020       NA    NA        NA
#>  9 0-4   Female Non Maori  2021       NA    NA        NA
#> 10 0-4   Female Non Maori  2022       NA    NA        NA
#> # ℹ 278 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>

## combined estimates and forecasts
mod |>
  forecast(labels = 2019:2024,
           include_estimates = TRUE)
#> # A tibble: 1,200 × 9
#>    age   sex    ethnicity  year injuries  popn .observed
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>
#>  1 0-4   Female Maori      2000       12 35830 0.000335 
#>  2 5-9   Female Maori      2000        6 35120 0.000171 
#>  3 10-14 Female Maori      2000        3 32830 0.0000914
#>  4 15-19 Female Maori      2000        6 27130 0.000221 
#>  5 20-24 Female Maori      2000        6 24380 0.000246 
#>  6 25-29 Female Maori      2000        6 24160 0.000248 
#>  7 30-34 Female Maori      2000       12 22560 0.000532 
#>  8 35-39 Female Maori      2000        3 22230 0.000135 
#>  9 40-44 Female Maori      2000        6 18130 0.000331 
#> 10 45-49 Female Maori      2000        6 13770 0.000436 
#> # ℹ 1,190 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>

## hyper-parameters
mod |>
  forecast(labels = 2019:2024,
           output = "components")
#> # A tibble: 6 × 4
#>   term  component level               .fitted
#>   <chr> <chr>     <chr>          <rdbl<1000>>
#> 1 year  effect    2019   -0.095 (-0.2, 0.022)
#> 2 year  effect    2020  -0.095 (-0.24, 0.051)
#> 3 year  effect    2021  -0.096 (-0.27, 0.083)
#> 4 year  effect    2022     -0.1 (-0.29, 0.11)
#> 5 year  effect    2023   -0.099 (-0.31, 0.14)
#> 6 year  effect    2024   -0.094 (-0.34, 0.15)

## forecast based on priors only
mod_unfitted <- mod_pois(injuries ~ age * sex + ethnicity + year,
                         data = injuries,
                         exposure = popn)
mod_unfitted |>
  forecast(labels = 2019:2024)
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 288 × 9
#>    age   sex    ethnicity  year injuries  popn .observed               .fitted
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>          <rdbl<1000>>
#>  1 0-4   Female Maori      2019       NA    NA        NA 0.56 (7.2e-06, 16475)
#>  2 0-4   Female Maori      2020       NA    NA        NA 0.49 (1.9e-05, 18877)
#>  3 0-4   Female Maori      2021       NA    NA        NA 0.53 (7.9e-06, 52802)
#>  4 0-4   Female Maori      2022       NA    NA        NA 0.51 (3.8e-06, 88882)
#>  5 0-4   Female Maori      2023       NA    NA        NA   0.49 (1e-05, 65947)
#>  6 0-4   Female Maori      2024       NA    NA        NA 0.57 (4.4e-06, 65858)
#>  7 0-4   Female Non Maori  2019       NA    NA        NA 0.56 (6.3e-06, 14505)
#>  8 0-4   Female Non Maori  2020       NA    NA        NA 0.68 (5.6e-06, 18252)
#>  9 0-4   Female Non Maori  2021       NA    NA        NA   0.52 (6e-06, 26505)
#> 10 0-4   Female Non Maori  2022       NA    NA        NA 0.52 (3.9e-06, 52483)
#> # ℹ 278 more rows
#> # ℹ 1 more variable: .expected <rdbl<1000>>