Generate Random Age or Age-Sex Profiles

Generate random age or age-sex profiles from an object of class "bage_ssvd". An object of class "bage_ssvd" holds results from an SVD decomposition of demographic data.

Usage

# S3 method for bage_ssvd
generate(x, n_draw = 20, n_comp = 5, joint = NULL, age_labels = NULL, ...)

Arguments

x

An object of class "bage_ssvd".

n_draw

Number of random draws to generate.

n_comp

The number of components to use when generating profiles. Default is 5. Maximum value is 10.

joint

Whether to use joint or independent SVDs for each sex/gender. If no value is supplied, an SVD for all sexes/genders combined is used.

age_labels

Age labels for the desired profile. If no labels are supplied, the most detailed profile available is used.

...

Not currently used.

Value

A tibble

Scaled SVDs currently implemented in bage

HMD Values from the Human Mortality Database.

TODO - add others

See also

Examples

## SVD for females and males combined
generate(HMD)
#> # A tibble: 2,220 × 3
#>     draw age   value
#>    <int> <chr> <dbl>
#>  1     1 0     -1.83
#>  2     1 1     -2.74
#>  3     1 2     -3.25
#>  4     1 3     -3.61
#>  5     1 4     -3.88
#>  6     1 5     -4.07
#>  7     1 6     -4.21
#>  8     1 7     -4.36
#>  9     1 8     -4.49
#> 10     1 9     -4.59
#> # ℹ 2,210 more rows

## separate SVDs for females and males
generate(HMD, joint = FALSE) 
#> # A tibble: 4,440 × 4
#>     draw sexgender age   value
#>    <int> <chr>     <chr> <dbl>
#>  1     1 Female    0     -3.44
#>  2     1 Female    1     -5.10
#>  3     1 Female    2     -5.52
#>  4     1 Female    3     -5.73
#>  5     1 Female    4     -5.88
#>  6     1 Female    5     -5.98
#>  7     1 Female    6     -6.08
#>  8     1 Female    7     -6.16
#>  9     1 Female    8     -6.23
#> 10     1 Female    9     -6.28
#> # ℹ 4,430 more rows

## specify age groups
labels <- poputils::age_labels(type = "lt", max = 60)
generate(HMD, age_labels = labels)
#> # A tibble: 280 × 3
#>     draw age   value
#>    <int> <chr> <dbl>
#>  1     1 0     -4.83
#>  2     1 1-4   -8.24
#>  3     1 5-9   -8.68
#>  4     1 10-14 -8.47
#>  5     1 15-19 -7.44
#>  6     1 20-24 -7.22
#>  7     1 25-29 -7.22
#>  8     1 30-34 -7.02
#>  9     1 35-39 -6.63
#> 10     1 40-44 -6.11
#> # ℹ 270 more rows