Draw from the Prior Predictive Distribution of a Model

Draw from the prior predictive distribution of a model, i.e., the Create Replicate Data

Usage

replicate_data(x, condition_on = NULL, n = 19)

Arguments

x

A fitted model, typically created by calling mod_pois(), mod_binom(), or mod_norm(), and then fit().

condition_on

Parameters to condition on. Either "meanpar" or "par". See details.

n

Number of replicate datasets to create. Default is 19.

Value

A tibble with the following structure:

.replicate	data

`"Original"	Original data supplied to mod_pois(), mod_binom(), mod_norm()

"Replicate 1"	Original data, except that actual outcome replaced by simulated values.

"Replicate 2"	Original data, except that actual outcome replaced by simulated values.

...	...

"Replicate <n>"	Original data, except that actual outcome replaced by simulated values.

Details

Use a fitted model to create replicate datasets, typically as a way of checking a model.

Use n draws from the posterior distribution for model parameters to generate n simulated datasets. If the model is working well, these simulated datasets should look similar to the actual dataset.

The condition_on argument

With Poisson and binomial models that include dispersion terms (which is the default), there are two options for constructing replicate data.

When condition_on is "par", the replicate data is created by (i) drawing values from the posterior distribution for rates or probabilities (the \(\gamma_i\) defined in mod_pois() and mod_binom()), and (ii) conditional on these rates or probabilities, drawing values for the outcome variable.

When condition_on is "meanpar", the replicate data is created by (i) drawing values from hyper-parameters governing the rates or probabilities (the \(\mu_i\) and \(\xi\) defined in mod_pois() and mod_binom()), then (ii) conditional on these hyper-parameters, drawing values for the rates or probabilties, and finally (iii) conditional on these rates or probabilities, drawing values for the outcome variable.

The default for condition_on is "meanpar". The "meanpar" option provides a more severe test for a model than the "par" option, since "par" values are weighted averages of the "meanpar" values and the original data.

As described in mod_norm(), normal models have a slightly different structure from Poisson and binomial models, and the distinction between "par" and "meanpar" does not apply.

See also

Examples

mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = 1) |>
  fit()

rep_data <- mod |>
  replicate_data()

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
rep_data |>
  group_by(.replicate) |>
  count(wt = injuries)
#> # A tibble: 20 × 2
#> # Groups:   .replicate [20]
#>    .replicate       n
#>    <fct>        <dbl>
#>  1 Original     21588
#>  2 Replicate 1  21513
#>  3 Replicate 2  21512
#>  4 Replicate 3  20871
#>  5 Replicate 4  21591
#>  6 Replicate 5  21277
#>  7 Replicate 6  21654
#>  8 Replicate 7  21356
#>  9 Replicate 8  21851
#> 10 Replicate 9  20542
#> 11 Replicate 10 21957
#> 12 Replicate 11 21789
#> 13 Replicate 12 21070
#> 14 Replicate 13 21624
#> 15 Replicate 14 21641
#> 16 Replicate 15 21377
#> 17 Replicate 16 21234
#> 18 Replicate 17 20684
#> 19 Replicate 18 21527
#> 20 Replicate 19 21875