Simulation study of a model

Use simulated data to assess the performance of and estimated model.

Usage

report_sim(
  mod_est,
  mod_sim = NULL,
  n_sim = 100,
  point_est_fun = c("median", "mean"),
  widths = c(0.5, 0.95),
  report_type = c("short", "long", "full"),
  n_core = 1
)

Arguments

mod_est

The model whose performance is being assessed. An object of class bage_mod.

mod_sim

The model used to generate the simulated data. If no value is supplied, mod_est is used.

n_sim

Number of sets of simulated data to use. Default is 100.

point_est_fun

Name of the function to use to calculate point estimates. The options are "mean" and "median". The default is "mean".

widths

Widths of credible intervals. A vector of values in the interval (0, 1]. Default is c(0.5, 0.95).

report_type

Amount of detail in return value. Options are "short" and "long". Default is "short".

n_core

Number of cores to use for parallel processing. If n_core is 1 (the default), no parallel processing is done.

Value

report_type is "short"

A tibble with the following columns:

component. Part of model. See Details. "par" is the rate, probability, or mean parameter from the likelihood.

vals_sim. Simulated value for parameter, averaged across all simulations and cells.

error_point_est. Point estimate minus simulation-true value, averaged across all simulations and cells.

cover. Actual proportion of simulation-true values that fall within each type of interval, averaged across all simulations and cells.

report_type is "long"

A tibble with the following columns:

component. Part of model. See components(). "par" is the rate, probability, or mean parameter from the likelihood.

term. Category within component.

level. Category within term.

vals_sim. Simulated values for parameter, stored in an rvec.

error_point_est. Point estimates minus simulation-true values, stored in an rvec.

cover. Actual proportions of simulation-true values falling within each type of interval, stored in an rvec.

Warning

The interface for report_sim is still experimental, and may change in future.

Comparisons included in report

When mod_est and mod_sim are identical, the report produced by report_sim() has the following components:

"hyper" Hyper-parameters from priors for intercept, main effects, and interactions.

"effect" Intercept, main effects, and interactions.

"meanpar" Expected value for rates probabilities. Poisson and binomial models only.

"disp" Dispersion term, if included in model.

"par" Rates, probabilities, or means.

When mod_est and mod_sim differ, the report omits "hyper", and "effect", since, for most model specifications, these terms are not comparable across models.

See also

Examples

## results random, so set seed
set.seed(0)

## make data - outcome variable (deaths here)
## needs to be present, but is not used
data <- data.frame(region = c("A", "B", "C", "D", "E"),
                   population = c(100, 200, 300, 400, 500),
                   deaths = NA)

## simulation with estimation model same as
## data-generating model
mod_est <- mod_pois(deaths ~ region,
                    data = data,
                    exposure = population) |>
  set_prior(`(Intercept)` ~ Known(0))
report_sim(mod_est = mod_est,
           n_sim = 10) ## in practice should use larger value
#> $components
#> # A tibble: 4 × 7
#>   term        component   .error .cover_50 .cover_95 .length_50 .length_95
#>   <chr>       <chr>        <dbl>     <dbl>     <dbl>      <dbl>      <dbl>
#> 1 (Intercept) effect    -0.0146       0.6        1        0.498       1.47
#> 2 region      effect     0.00316      0.68       1        0.920       2.63
#> 3 region      hyper      0.0211       0.4        0.9      0.574       1.89
#> 4 disp        disp       0.0405       0.5        0.9      0.798       2.78
#> 
#> $augment
#> # A tibble: 2 × 7
#>   .var      .observed  .error .cover_50 .cover_95 .length_50 .length_95
#>   <chr>         <dbl>   <dbl>     <dbl>     <dbl>      <dbl>      <dbl>
#> 1 .fitted        4.41 -0.0108      0.58      0.98      0.116      0.330
#> 2 .expected      4.41 -0.521       0.66      0.98      3.08      15.0  
#> 

## simulation with estimation model different
## from data-generating model
mod_sim <- mod_est |>
  set_prior(region ~ N(s = 2))
report_sim(mod_est = mod_est,
           mod_sim = mod_sim,
           n_sim = 10)
#> $components
#> # A tibble: 4 × 7
#>   term        component   .error .cover_50 .cover_95 .length_50 .length_95
#>   <chr>       <chr>        <dbl>     <dbl>     <dbl>      <dbl>      <dbl>
#> 1 (Intercept) effect     0.114        0.5       0.9       0.547       1.64
#> 2 region      effect     0.00131      0.56      0.96      1.04        3.00
#> 3 region      hyper     -0.516        0.2       0.8       0.628       2.01
#> 4 disp        disp       0.127        0.5       0.8       1.05        4.64
#> 
#> $augment
#> # A tibble: 2 × 7
#>   .var      .observed  .error .cover_50 .cover_95 .length_50 .length_95
#>   <chr>         <dbl>   <dbl>     <dbl>     <dbl>      <dbl>      <dbl>
#> 1 .fitted        157.   0.207      0.58      0.96      0.391       1.08
#> 2 .expected      157. -12.3        0.6       0.98    131.        730.  
#>