Create components needed by a scaled_svd object

Use the Singular Value Decomposition (SVD) to construct a parsimonious representation of a set of rates, probabilities, means, or other values.

Usage

scaled_svd_comp(
  x,
  n = 10,
  transform = c("log", "logit", "none"),
  center = TRUE
)

Arguments

x

A matrix with value such as rates, probabilities, or means.

n

Number of components of the SVD to use. Defaults to 10.

transform

"log", "logit", or "none". Defaults to "log".

center

Whether to center the transformed version of x before carrying out the SVD.

Value

A named list with two elements:

matrix, a numeric matrix

offset, a numeric vector

Details

`scaled_svd_comp() typically proceeds as follows:

transform values in matrix x (eg take logs)

center each column of x

carry out a SVD on the transformed version of x

centre and scale the results from the SVD to produce matrix matrix and vector offset.

If

\(X\) is a the matrix of transformed values of x

\(A\) is the matrix called matrix,

\(b\) is the vector called offset, and

\(\beta\) is a vector of standard normal variates,

and

$$v = A \beta + b$$

then \(v\) should look like a randomly-selected column from \(X\).

Matrix x typically has age along the rows, and some combination of classification variables, such as country and time, along the columns. One exception is when the SVD is used to capture the relationship between female and male rates, in which case rows are formed by interacting sex and age. See below for an example.

The number of components used by the SVD (and hence the number of columns of the tranform matrix) is governed by argument n. The default is 5.

When scale is "log" or "logit", scaled_svd_comp() converts any 0s in x to values just above 0 before applying the log or logit function. When scale is "logit", scaled_svd_comp() also converts any 1s to values just below 1`.

Centering is appropriate when modelling the shape of, for instance, age profiles, rather than the absolute level of the profiles.

See also

Examples

x <- matrix(rgamma(n = 50, shape = 1), nrow = 5, ncol = 10)
x
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#> [1,] 0.2772443 2.3161305 0.6456835 0.9560931 0.4746108 1.0243892 1.2981059
#> [2,] 1.4996090 2.6847973 0.6805910 0.3921152 0.4965372 0.8580878 1.3076182
#> [3,] 1.0076971 0.4475368 0.1089154 0.6146813 0.1870204 0.2077955 0.8426195
#> [4,] 1.0373604 0.7849718 0.6735251 0.3510826 1.3277591 0.4096957 0.4559454
#> [5,] 1.2705326 3.0140234 0.5384111 0.1610642 0.5427547 0.5758258 0.1786242
#>            [,8]      [,9]      [,10]
#> [1,] 2.20486571 0.3482980 4.45094629
#> [2,] 0.18297872 3.3571084 0.89208882
#> [3,] 1.57042732 4.2511971 2.06341276
#> [4,] 3.11024256 0.1768955 1.27038410
#> [5,] 0.03239837 3.9062966 0.03056139
scaled_svd_comp(x, n = 3)
#> $matrix
#> 5 x 3 sparse Matrix of class "dgCMatrix"
#>                                      
#> [1,] -0.8550770 -0.2658565 -0.3366196
#> [2,]  0.3347385  0.1235422 -0.3891796
#> [3,] -0.3616849  0.8750007  0.2196997
#> [4,] -0.6352024 -0.5386675  0.3930293
#> [5,]  1.5172257 -0.1940189  0.1130702
#> 
#> $offset
#> [1]  0.46095851  0.13084765 -0.03602179 -0.03721707 -0.51856729
#> 
scaled_svd_comp(x, n = 3, transform = "none")
#> $matrix
#> 5 x 3 sparse Matrix of class "dgCMatrix"
#>                                        
#> [1,] -1.0786666 -0.53826763  0.09317444
#> [2,]  0.6051643 -0.16639263  0.22511270
#> [3,]  0.2902318  0.01401833 -0.76430364
#> [4,] -0.7530706  0.66926342  0.13704436
#> [5,]  0.9363410  0.02137851  0.30897215
#> 
#> $offset
#> [1]  0.26462878  0.01942493 -0.01670553 -0.19628516 -0.07106303
#>