Set the number of draws

Specify the number of draws from the posterior distribution to be used in model output. A newly-created bage_mod object has an n_draw value of 1000. Higher values may be appropriate for characterising the tails of distributions, or for publication-quality graphics and summaries.

Usage

set_n_draw(mod, n_draw = 1000L)

Arguments

mod

A bage_mod object, typically created with mod_pois(), mod_binom(), or mod_norm().

n_draw

Number of draws.

Value

A bage_mod object

Details

The value of n_draw does not affect model fitting: it only affects posterior summaries.

See also

Examples

mod <- mod_pois(injuries ~ age:sex + ethnicity + year,
                data = injuries,
                exposure = popn)
mod
#> -- Unfitted Poisson model --
#> 
#>    injuries ~ age:sex + ethnicity + year
#> 
#> (Intercept) ~ NFix()
#>   ethnicity ~ NFix()
#>        year ~ RW()
#>     age:sex ~ ERW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>   var_sexgender: sex
#>        var_time: year
#>          n_draw: 1000

mod |>
  set_n_draw(n_draw = 5000)
#> -- Unfitted Poisson model --
#> 
#>    injuries ~ age:sex + ethnicity + year
#> 
#> (Intercept) ~ NFix()
#>   ethnicity ~ NFix()
#>        year ~ RW()
#>     age:sex ~ ERW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>   var_sexgender: sex
#>        var_time: year
#>          n_draw: 5000