Set a sex or gender variable

Specify which variable (if any) represents sex or gender. Functions mod_pois(), mod_binom(), and mod_norm() try to find a sex or gender variable using variable names, but do not always get it right.

Usage

set_var_sexgender(mod, name)

Arguments

mod

A bage_mod object, typically created with mod_pois(), mod_binom(), or mod_norm().

name

The name of the sex or gender variable.

Value

A "bage_mod" object

Details

In an R formula, a 'variable' is different from a 'term'. For instance,

~ gender + region + gender:region

contains variables gender and region, and terms gender, region, and gender:region.

If the mod argument to set_var_sexgender is a fitted model, then set_var_sexgender 'unfits' the model, by deleting existing estimates.

See also

Examples

## rename 'sex' variable to something unexpected
injuries2 <- injuries
injuries2$biological_sex <- injuries2$sex

## mod_pois does not recognize sex variable
mod <- mod_pois(injuries ~ age * biological_sex + year,
                data = injuries2,
                exposure = popn)
mod
#> -- Unfitted Poisson model --
#> 
#>           injuries ~ age * biological_sex + year
#> 
#>        (Intercept) ~ NFix()
#>                age ~ RW()
#>     biological_sex ~ NFix()
#>               year ~ RW()
#> age:biological_sex ~ ERW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>        var_time: year
#>          n_draw: 1000

## so we set the sex variable explicitly
mod |>
  set_var_sexgender(name = "biological_sex")
#> -- Unfitted Poisson model --
#> 
#>           injuries ~ age * biological_sex + year
#> 
#>        (Intercept) ~ NFix()
#>                age ~ RW()
#>     biological_sex ~ NFix()
#>               year ~ RW()
#> age:biological_sex ~ ERW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>   var_sexgender: biological_sex
#>        var_time: year
#>          n_draw: 1000