Create an Object to Hold Data from a Scaled SVD

Create an object of class "bage_ssvd" to hold results from a scaled Singular Value Decomposition (SVD).

Usage

ssvd(data)

Arguments

data

A data frame. See Details for description.

Value

An object of class "bage_ssvd".

Details

data has the following columns:

type Type of decomposition. Choices are "total", "joint", and "indep".

labels_age Age labels for individual rows of matrices within matrix and individual elements of vectors within offset.

labels_sexgender Sex/gender labels for individual rows of matrices within matrix and individual elements of vectors within offset, or NULL. NULL when sexgender is "total", since in this case results average across sexes/genders.

matrix List column of sparse matrices created by ssvd_comp(). Must have rownames. Must not have NAs. When type is "total" or "joint", each matrix has 10 columns. When "type" is "indep", each matrix has 20 columns.

offset List column of vectors created by ssvd_comp(). Must have names, which are identical to the rownames of the corresponding element of matrix.