Create a Scaled SVD Object from a Human Mortality Database File

Create an object of class "bage_ssvd" from a zipped file downloaded from the Human Mortality Database.

Usage

ssvd_hmd(zipfile, transform = c("log", "logit", "none"))

Arguments

zipfile

The name of a zipped file downloaded from the Human Mortality Database. Any path name that can be handled by utils::unzip().

transform

"log", "logit", or "none". Defaults to "log".

Value

An object of class "bage_ssvd"

Warning

This function, an associated functions, will probably be moved into a separate package in future.

Usage

Step 1: Download data

Register or log in at the Human Mortality Database, and go to page Downloading the HMD in zipped data files. Go to the "Previous Versions" table at the bottom of the page, and download a file from the "Statistics" column, eg file

https://www.mortality.org/File/Download/hmd.v6/zip/all_hmd/hmd_statistics_20240226.zip

To download a file, you will need to register with the Human Mortality Databse and log in.

Step 2: Call function 'svd_hmd'

Put the file into the working directory, and supply the name of the file to function svd_hmd(), eg

MyHMD <- svd_hmd("hmd_statistcs_20240226")

Step 3: Use object to create a prior

mod |>
  set_prior(age ~ SVD(MyHMD))

See also