Exchangeable SVD Prior for Interactions Involving Age and Sex/Gender

Specify a Singular Value Decomposition (SVD) prior for an interaction involving age, sex/gender, and at least one other dimension. Age-sex profiles are "exchangeable", in that profiles for all combinations of the non-age, non-sex variables are drawn from the same distribution. The age-sex profiles are assumed to look like they were drawn at random from an external demographic database. For instance,

set_prior(mod, age:sex ~ ESVDS(HMD))

specifies that age-sex profiles should look like they was drawn from the Human Mortality Database.

Usage

ESVDS(ssvd, n = 5, joint = FALSE)

Arguments

ssvd

Object of class "bage_ssvd" holding results from a scaled SVD. See below for current options.

n

Number of components from scaled SVD to use in modelling. The default is half the number of components of ssvd.

joint

Whether to use combined or separate SVDs. Default is FALSE. See below for details.

Value

An object of class "bage_prior_esvd".

Mathematical details

Let \(V\) denote the number of age groups, \(S\) the number of sexes/genders, and \(U\) the number of combinations of non-age, non-sex variables. When joint is FALSE,

$$\pmb{\beta}_{us} = \pmb{F}_s \pmb{\alpha}_{us} + \pmb{g}_s, \quad, u = 1, \cdots, U, s = 1, \cdots S$$

where \(\pmb{\beta}_{us}\), \(\pmb{\alpha}_{us}\), and \(\pmb{g}_{us}\) each have \(V\) elements, and \(\pmb{F}\) has \(V\) rows.

When joint is TRUE,

$$\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g}$$

where

where \(\pmb{\beta}_u\), \(\pmb{\alpha}_u\), and \(\pmb{g}\) each have \(V \times S\) elements, and \(\pmb{F}\) has \(V \times S\) rows.

\(\pmb{F}_s\), \(\pmb{g}_s\), \(\pmb{F}\), and \(\pmb{g}\), are all estimated externally, using estimates from a demographic database, and are treated as fixed and known.

The elements of \(\pmb{\alpha}_{us}\) and \(\pmb{\alpha}_u\) have prior

$$\alpha_{uk} \sim \text{N}(0, 1)$$

For a description of the construction of \(\pmb{F}_s\), \(\pmb{F}\), \(\pmb{g}_s\), and \(\pmb{g}\), see TODO - REFERENCE TO VIGNETTE.

Joint or independent SVDs

To possible ways of extracting patterns from age-sex-specific data are

carry out separate SVDs on separate datasets for each sex/gender; or

carry out a single SVD on dataset that has separate entries for each sex/gender.

Option 1 is more flexible. Option 2 is more robust to sampling or measurement errors. Option 1 is obtained by setting the joint argument to FALSE. Option 2 is obtained by setting the joint argument to TRUE. The default is FALSE.

Scaled SVDs currently implemented in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

See also

Examples

ESVDS(HMD)
#> $i_prior
#> [1] 14
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> [1] 5
#> 
#> $specific$joint
#> [1] FALSE
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_esvd" "bage_prior"     
ESVDS(HMD, joint = TRUE)
#> $i_prior
#> [1] 14
#> 
#> $const
#> [1] 0
#> 
#> $specific
#> $specific$ssvd
#> 
#> ------------------------------------------------------------ 
#> <Object of class "bage_ssvd">
#> 
#> sex/gender labels:
#>      Female 
#>      Male 
#> 
#> age labels:
#>      0, 1, 2, 3, ..., 58, 59, 60+ 
#>      0, 1, 2, 3, ..., 63, 64, 65+ 
#>      0, 1, 2, 3, ..., 68, 69, 70+ 
#>      0, 1, 2, 3, ..., 73, 74, 75+ 
#>      0, 1, 2, 3, ..., 78, 79, 80+ 
#>      0, 1, 2, 3, ..., 83, 84, 85+ 
#>      0, 1, 2, 3, ..., 88, 89, 90+ 
#>      0, 1, 2, 3, ..., 93, 94, 95+ 
#>      0, 1, 2, 3, ..., 98, 99, 100+ 
#>      0, 1, 2, 3, ..., 103, 104, 105+ 
#>      0, 1, 2, 3, ..., 108, 109, 110+ 
#>      0, 1-4, 5-9, 10-14, ..., 50-54, 55-59, 60+ 
#>      0, 1-4, 5-9, 10-14, ..., 55-59, 60-64, 65+ 
#>      0, 1-4, 5-9, 10-14, ..., 60-64, 65-69, 70+ 
#>      0, 1-4, 5-9, 10-14, ..., 65-69, 70-74, 75+ 
#>      0, 1-4, 5-9, 10-14, ..., 70-74, 75-79, 80+ 
#>      0, 1-4, 5-9, 10-14, ..., 75-79, 80-84, 85+ 
#>      0, 1-4, 5-9, 10-14, ..., 80-84, 85-89, 90+ 
#>      0, 1-4, 5-9, 10-14, ..., 85-89, 90-94, 95+ 
#>      0, 1-4, 5-9, 10-14, ..., 90-94, 95-99, 100+ 
#>      0, 1-4, 5-9, 10-14, ..., 95-99, 100-104, 105+ 
#>      0, 1-4, 5-9, 10-14, ..., 100-104, 105-109, 110+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 50-54, 55-59, 60+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 55-59, 60-64, 65+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 60-64, 65-69, 70+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 65-69, 70-74, 75+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 70-74, 75-79, 80+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 75-79, 80-84, 85+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 80-84, 85-89, 90+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 85-89, 90-94, 95+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 90-94, 95-99, 100+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 95-99, 100-104, 105+ 
#>      0-4, 5-9, 10-14, 15-19, ..., 100-104, 105-109, 110+ 
#> ------------------------------------------------------------ 
#> 
#> $specific$nm_ssvd
#> [1] "HMD"
#> 
#> $specific$n
#> [1] 5
#> 
#> $specific$joint
#> [1] TRUE
#> 
#> 
#> attr(,"class")
#> [1] "bage_prior_esvd" "bage_prior"