SVD Prior for Age-Sex or Age-Gender Interaction

A Singular Value Decomposition (SVD) prior for an interaction involving age-sex profiles or age-gender profiles.

Usage

SVDS(ssvd, n_comp = NULL, joint = FALSE)

Arguments

ssvd

Object of class "bage_ssvd" holding a scaled SVD. See below for scaled SVDs of databases currently available in bage.

n_comp

Number of components from scaled SVD to use in modelling. The default is half the number of components of ssvd.

joint

Whether to use combined or separate SVDs. Default is FALSE. See below for details.

Value

An object of class "bage_prior_svd".

Details

A SVDS() prior assumes that the age-sex profile or age-gender profile for the quantity being modelled looks like it was drawn at random from an external demographic database. For instance, the prior obtained via

SVDS(HMD)

assumes that age-sex profiles look like they were obtained from the Human Mortality Database.

If SVDS() is used with an interaction between age, sex/gender, and one or more other variables, then separate age-sex or age-gender profiles are constructed within each combination of the other variables.

Age profiles with no sex or gender dimension should be modelled using SVD() rather than SVDS().

Joint or independent SVDs

Two possible ways of extracting patterns from age-sex-specific data are

carry out separate SVDs on separate datasets for each sex/gender; or

carry out a single SVD on dataset that has separate entries for each sex/gender.

Option 1 is more flexible. Option 2 is more robust to sampling or measurement errors. Option 1 is obtained by setting the joint argument to FALSE. Option 2 is obtained by setting the joint argument to TRUE. The default is FALSE.

Mathematical details

Case 1: Interaction involving age and sex or gender, but no other variables

When joint is FALSE,

$$\pmb{\beta}_{s} = \pmb{F}_s \pmb{\alpha}_{s} + \pmb{g}_s,$$

and when joint is TRUE,

$$\pmb{\beta} = \pmb{F} \pmb{\alpha} + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction involving age and sex/gender;

\(\pmb{\beta}_{s}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\);

\(J\) is the number of elements in \(\pmb{\beta}\);

\(S\) is the number of sexes/genders;

\(n\) is n_comp;

\(\pmb{F}_s\) is a known \((J/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(J/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(J \times n\) matrix, with values for all sexes/genders; and

\(\pmb{g}\) is a known vector with \(J\) elements, with values for all sexes/genders.

The elements of \(\pmb{\alpha}_s\) and \(\pmb{\alpha}\) have prior $$\alpha_k \sim \text{N}(0, 1).$$

Case 2: Interaction involving age and sex/gender, and one or more other variables

When joint is FALSE,

$$\pmb{\beta}_{u,s} = \pmb{F}_s \pmb{\alpha}_{u,s} + \pmb{g}_s,$$

and when joint is TRUE,

$$\pmb{\beta}_u = \pmb{F} \pmb{\alpha}_u + \pmb{g},$$

where

\(\pmb{\beta}\) is an interaction involving sex/gender;

\(\pmb{\beta}_{u,s}\) is a subvector of \(\pmb{\beta}\), holding values for sex/gender \(s\) for the \(u\)th combination of the other variables;

\(V\) is the number of elements in \(\pmb{\beta}_u\);

\(S\) is the number of sexes/genders;

\(n\) is n_comp;

\(\pmb{F}_s\) is a known \((V/S) \times n\) matrix, specific to sex/gender \(s\);

\(\pmb{g}_s\) is a known vector with \(V/S\) elements, specific to sex/gender \(s\);

\(\pmb{F}\) is a known \(V \times n\) matrix, with values for all sexes/genders; and

\(\pmb{g}\) is a known vector with \(V\) elements, with values for all sexes/genders.

The elements of \(\pmb{\alpha}_{u,s}\) and \(\pmb{\alpha}_u\) have prior $$\alpha_k \sim \text{N}(0, 1).$$

Scaled SVDs of demographic databases in bage

HMD Mortality rates from the Human Mortality Database.

LFP Labour forcce participation rates from the OECD.

References

For details of the construction of scaled SVDS see the vignette here.

See also

Examples

SVDS(HMD)
#>   SVDS(HMD) 
#>     ssvd: HMD
#>   n_comp: 5
#>    joint: FALSE
SVDS(HMD, joint = TRUE)
#>   SVDS(HMD,joint=TRUE) 
#>     ssvd: HMD
#>   n_comp: 5
#>    joint: TRUE