Data and Values from a Model

Extract data and values from a model object with class "bage_mod". The returne value is a tibble containing the original data, plus draws from the posterior distribution if the model has been fitted, or the prior distribution if it has not.

Usage

# S3 method for bage_mod
augment(x, quiet = FALSE, ...)

Arguments

x

A bage_mod object.

quiet

Whether to suppress messages. Default is FALSE.

...

Unused. Included for generic consistency only.

Value

A tibble.

Details

The return value contains the following columns:

.observed 'Direct' estimates of rates or probabilities, ie counts divided by exposure/size. (Poisson and binomial models only.)

.fitted Draws of rates, probabilities, or means.

expected Draws of expected values for rates or probabilities. (Inly in Poisson that include exposure, or in binomial models.)

Uncertain quantities are represented using rvecs.

See also

Examples

## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> augment()
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 912 × 9
#>    age   sex    ethnicity  year            injuries  popn           .observed
#>    <fct> <chr>  <chr>     <int>        <rdbl<1000>> <int>        <rdbl<1000>>
#>  1 0-4   Female Maori      2000 26091 (12, 1.5e+07) 35830 0.73 (0.00033, 414)
#>  2 5-9   Female Maori      2000 25769 (16, 1.1e+07) 35120 0.73 (0.00046, 317)
#>  3 10-14 Female Maori      2000 23152 (17, 1.1e+07) 32830 0.71 (0.00051, 336)
#>  4 15-19 Female Maori      2000 20067 (22, 8094501) 27130 0.74 (0.00081, 298)
#>  5 20-24 Female Maori      2000 16064 (33, 6118622) 24380  0.66 (0.0014, 251)
#>  6 25-29 Female Maori      2000   17744 (18, 4e+06) 24160 0.73 (0.00074, 167)
#>  7 30-34 Female Maori      2000   18134 (17, 6e+06) 22560  0.8 (0.00075, 266)
#>  8 35-39 Female Maori      2000 17766 (15, 8560685) 22230  0.8 (0.00067, 385)
#>  9 40-44 Female Maori      2000 14416 (13, 7353526) 18130  0.8 (0.00072, 406)
#> 10 45-49 Female Maori      2000   9256 (9, 3690240) 13770 0.67 (0.00065, 268)
#> # ℹ 902 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> augment()
#> # A tibble: 912 × 9
#>    age   sex    ethnicity  year injuries  popn .observed
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>
#>  1 0-4   Female Maori      2000       12 35830 0.000335 
#>  2 5-9   Female Maori      2000        6 35120 0.000171 
#>  3 10-14 Female Maori      2000        3 32830 0.0000914
#>  4 15-19 Female Maori      2000        6 27130 0.000221 
#>  5 20-24 Female Maori      2000        6 24380 0.000246 
#>  6 25-29 Female Maori      2000        6 24160 0.000248 
#>  7 30-34 Female Maori      2000       12 22560 0.000532 
#>  8 35-39 Female Maori      2000        3 22230 0.000135 
#>  9 40-44 Female Maori      2000        6 18130 0.000331 
#> 10 45-49 Female Maori      2000        6 13770 0.000436 
#> # ℹ 902 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>