Extract Components from a Model

Extract batches of hyper-parameters from a model object with class "bage_mod". tibble draws from the posterior distribution if the model has been fitted, or the prior distribution if it has not.

Usage

# S3 method for bage_mod
components(object, quiet = FALSE, ...)

Arguments

object

A bage_mod object.

quiet

Whether to suppress messages. Default is FALSE.

...

Unused. Included for generic consistency only.

Value

A tibble

Details

The return value contains the following columns:

term Model term that the hyper-parameter belongs to.

component Component within term.

level Element within component .

.fitted An rvec containing draws from the posterior distribution.

See also

Examples

## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## look at prior distribution
mod |> components()
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 37 × 4
#>    term        component level                   .fitted
#>    <chr>       <chr>     <chr>              <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)            0 (0, 0)
#>  2 age         effect    0-4           0.034 (-4.2, 4.6)
#>  3 age         effect    5-9           0.031 (-3.8, 3.6)
#>  4 age         effect    10-14        0.0013 (-3.2, 2.9)
#>  5 age         effect    15-19       -0.0013 (-2.6, 2.3)
#>  6 age         effect    20-24        -0.002 (-2.6, 2.2)
#>  7 age         effect    25-29         -0.0078 (-2, 2.1)
#>  8 age         effect    30-34       -0.0075 (-1.9, 2.3)
#>  9 age         effect    35-39        -0.011 (-1.9, 2.4)
#> 10 age         effect    40-44       -0.0011 (-2.6, 2.7)
#> # ℹ 27 more rows

## fit model
mod <- mod |>
  fit()

## look at posterior distribution
mod |> components() ## posterior distribution
#> # A tibble: 37 × 4
#>    term        component level                    .fitted
#>    <chr>       <chr>     <chr>               <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)    -8.2 (-8.2, -8.2)
#>  2 age         effect    0-4         -0.36 (-0.47, -0.25)
#>  3 age         effect    5-9            -1.8 (-1.9, -1.6)
#>  4 age         effect    10-14          -1.3 (-1.4, -1.2)
#>  5 age         effect    15-19          0.47 (0.38, 0.56)
#>  6 age         effect    20-24          0.57 (0.47, 0.66)
#>  7 age         effect    25-29          0.46 (0.36, 0.56)
#>  8 age         effect    30-34          0.41 (0.31, 0.51)
#>  9 age         effect    35-39          0.34 (0.24, 0.44)
#> 10 age         effect    40-44          0.34 (0.24, 0.45)
#> # ℹ 27 more rows