Extract Values for Hyper-Parameters

Extract values for hyper-parameters from a model object. Includes values for main effects and interactions, and values for any disperion or variance terms.

Usage

# S3 method for class 'bage_mod'
components(object, standardize = TRUE, quiet = FALSE, ...)

Arguments

object

A bage_mod object, typically created with mod_pois(), mod_binom(), or mod_norm().

standardize

Whether to standardize parameter estimates. See below for details. Default is TRUE.

quiet

Whether to suppress messages. Default is FALSE.

...

Unused. Included for generic consistency only.

Value

A tibble with four columns columns:

The return value contains the following columns:

term Model term that the hyper-parameter belongs to.

component Component within term.

level Element within component .

.fitted An rvec containing draws from the posterior distribution.

Fitted vs unfitted models

components() is typically called on a fitted model. In this case, the modelled values are draws from the joint posterior distribution for the hyper-parameters in the model.

components() can, however, be called on an unfitted model. In this case, the modelled values are draws from the joint prior distribution. In other words, the modelled values are informed by model priors, and by any exposure, size, or weights argument in the model, but not by the observed outcomes.

Standardizing estimates

In many models, the sum of the intercept, main effects, and interactions is well-identified by the data, but the values of the intercept, main effects, and interactions are not. For instance, consider a model containing the lines

$$y_i \sim \text{N}(\mu_i, \sigma^2)$$ $$\mu_i = \beta^{(0)} + \beta_i^{(1)}.$$

If we add 1 to \(\beta^{(0)}\) and subtract 1 from each of the \(\beta_i^{(1)}\), then the \(\mu_i\) are unchanged, and the model's fit to the \(y_i\) is unchanged. This indeterminacy does not affect our ability to estimate the \(\mu_i\), but does complicate the interpretation of \(\beta^{(0)}\) and the \(\beta_i^{(1)}\).

A common strategy for dealing with poorly- identified intercepts, main effects, and interactions is to post-process the estimates, so that all terms other than the intercept, are constrained to sum to zero. By default, this is what bage does. For a description of the algorithm used by bage, see vignette("vig2_math").

For unstandardized estimates, set standardize to FALSE.

Batches of parameters, such as seasonal effects, that have the same structure as main effects and interactions have the same problems of indeterminacy. By default, these paramters are also standardized.

See also

Examples

## specify model
mod <- mod_pois(injuries ~ age + sex + year,
                data = injuries,
                exposure = popn)

## extract prior distribution
## of hyper-parameters
mod |>
  components()
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 37 × 4
#>    term        component level                   .fitted
#>    <chr>       <chr>     <chr>              <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)   -0.11 (-7.5, 7.3)
#>  2 age         effect    0-4           0.019 (-3.6, 3.7)
#>  3 age         effect    5-9           0.011 (-3.5, 3.2)
#>  4 age         effect    10-14         0.00098 (-3, 2.6)
#>  5 age         effect    15-19       -0.0039 (-2.9, 2.6)
#>  6 age         effect    20-24       -0.0013 (-2.4, 2.5)
#>  7 age         effect    25-29       0.00025 (-2.2, 2.3)
#>  8 age         effect    30-34         -0.0017 (-2.3, 2)
#>  9 age         effect    35-39       -0.0044 (-2.3, 2.1)
#> 10 age         effect    40-44        -0.018 (-2.5, 2.5)
#> # ℹ 27 more rows

## fit model
mod <- mod |>
  fit()

## extract posterior distribution
## of hyper-parameters
mod |>
  components()
#> # A tibble: 37 × 4
#>    term        component level                    .fitted
#>    <chr>       <chr>     <chr>               <rdbl<1000>>
#>  1 (Intercept) effect    (Intercept)    -8.2 (-8.2, -8.2)
#>  2 age         effect    0-4         -0.36 (-0.47, -0.25)
#>  3 age         effect    5-9            -1.8 (-1.9, -1.6)
#>  4 age         effect    10-14          -1.3 (-1.4, -1.2)
#>  5 age         effect    15-19          0.47 (0.37, 0.56)
#>  6 age         effect    20-24          0.56 (0.46, 0.66)
#>  7 age         effect    25-29          0.46 (0.36, 0.55)
#>  8 age         effect    30-34          0.41 (0.31, 0.51)
#>  9 age         effect    35-39          0.34 (0.24, 0.45)
#> 10 age         effect    40-44          0.34 (0.25, 0.45)
#> # ℹ 27 more rows