Use a Model to Make a Forecast

Forecast rates, probabilities, means, and other model parameters.

Usage

# S3 method for class 'bage_mod'
forecast(
  object,
  output = c("augment", "components"),
  include_estimates = FALSE,
  standardize = c("terms", "anova", "none"),
  labels,
  ...
)

Arguments

object

A bage_mod object, typically created with mod_pois(), mod_binom(), or mod_norm().

output

Type of output returned

include_estimates

Whether to include historical estimates along with the forecasts. Default is FALSE.

standardize

Standardization method: one of "terms", "anova", or "none". The default is "terms". See below for details.

labels

Labels for future values.

...

Not currently used.

Value

A tibble.

How the forecasts are constructed

Internally, the steps involved in a forecast are:

Forecast time-varying main effects and interactions, e.g. a time main effect, or an age-time interaction.

Combine forecasts for the time-varying main effects and interactions with non-time-varying parameters, e.g. age effects or dispersion.

Use the combined parameters to generate values for rates, probabilities or means.

vignette("vig2_math") has the technical details.

Output

When output is "augment" (the default), the return value from forecast() looks like output from function augment(). When output is "components", the return value looks like output from components().

When include_estimates is FALSE (the default), the output of forecast() excludes values for time-varying parameters for the period covered by the data. When include_estimates is TRUE, the output includes these values. Setting include_estimates to TRUE can be helpful when creating graphs that combine estimates and forecasts.

Standardization

The standardization used by forecast() is equivalent to the standardization applied by components(), except that values for forecasted terms are are shifted so that they line up with the values for estimates.

Fitted and unfitted models

forecast() is typically used with a fitted model, i.e. a model in which parameter values have been estimated from the data. The resulting forecasts reflect data and priors.

forecast() can, however, be used with an unfitted model. In this case, the forecasts are based entirely on the priors. See below for an example. Experimenting with forecasts based entirely on the priors can be helpful for choosing an appropriate model.

Warning

The interface for forecast() has not been finalised.

See also

Examples

## specify and fit model
mod <- mod_pois(injuries ~ age * sex + ethnicity + year,
                data = injuries,
                exposure = popn) |>
  fit(mod)
mod
#> -- Fitted Poisson model --
#> 
#>    injuries ~ age * sex + ethnicity + year
#> 
#> (Intercept) ~ NFix()
#>         age ~ RW()
#>         sex ~ NFix()
#>   ethnicity ~ NFix()
#>        year ~ RW()
#>     age:sex ~ RW()
#> 
#>      dispersion: mean=1
#>        exposure: popn
#>         var_age: age
#>   var_sexgender: sex
#>        var_time: year
#>          n_draw: 1000

## forecasts
mod |>
  forecast(labels = 2019:2024)
#> # A tibble: 288 × 9
#>    age   sex   ethnicity  year injuries  popn .observed                  .fitted
#>    <fct> <chr> <chr>     <int>    <int> <int>     <dbl>             <rdbl<1000>>
#>  1 0-4   Fema… Maori      2019       NA    NA        NA 2e-04 (0.00014, 0.00027)
#>  2 0-4   Fema… Maori      2020       NA    NA        NA 2e-04 (0.00014, 0.00028)
#>  3 0-4   Fema… Maori      2021       NA    NA        NA 2e-04 (0.00014, 0.00028)
#>  4 0-4   Fema… Maori      2022       NA    NA        NA 2e-04 (0.00013, 0.00029)
#>  5 0-4   Fema… Maori      2023       NA    NA        NA 2e-04 (0.00013, 0.00029)
#>  6 0-4   Fema… Maori      2024       NA    NA        NA 2e-04 (0.00013, 0.00028)
#>  7 0-4   Fema… Non Maori  2019       NA    NA        NA 1e-04 (7.5e-05, 0.00014)
#>  8 0-4   Fema… Non Maori  2020       NA    NA        NA 1e-04 (7.6e-05, 0.00015)
#>  9 0-4   Fema… Non Maori  2021       NA    NA        NA 1e-04 (7.1e-05, 0.00015)
#> 10 0-4   Fema… Non Maori  2022       NA    NA        NA 1e-04 (7.3e-05, 0.00015)
#> # ℹ 278 more rows
#> # ℹ 1 more variable: .expected <rdbl<1000>>

## combined estimates and forecasts
mod |>
  forecast(labels = 2019:2024,
           include_estimates = TRUE)
#> # A tibble: 1,200 × 9
#>    age   sex    ethnicity  year injuries  popn .observed
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>
#>  1 0-4   Female Maori      2000       12 35830 0.000335 
#>  2 5-9   Female Maori      2000        6 35120 0.000171 
#>  3 10-14 Female Maori      2000        3 32830 0.0000914
#>  4 15-19 Female Maori      2000        6 27130 0.000221 
#>  5 20-24 Female Maori      2000        6 24380 0.000246 
#>  6 25-29 Female Maori      2000        6 24160 0.000248 
#>  7 30-34 Female Maori      2000       12 22560 0.000532 
#>  8 35-39 Female Maori      2000        3 22230 0.000135 
#>  9 40-44 Female Maori      2000        6 18130 0.000331 
#> 10 45-49 Female Maori      2000        6 13770 0.000436 
#> # ℹ 1,190 more rows
#> # ℹ 2 more variables: .fitted <rdbl<1000>>, .expected <rdbl<1000>>

## hyper-parameters
mod |>
  forecast(labels = 2019:2024,
           output = "components")
#> # A tibble: 6 × 4
#>   term  component level               .fitted
#>   <chr> <chr>     <chr>          <rdbl<1000>>
#> 1 year  effect    2019   -0.096 (-0.2, 0.018)
#> 2 year  effect    2020  -0.094 (-0.22, 0.044)
#> 3 year  effect    2021  -0.093 (-0.27, 0.081)
#> 4 year  effect    2022  -0.091 (-0.29, 0.095)
#> 5 year  effect    2023    -0.096 (-0.3, 0.13)
#> 6 year  effect    2024   -0.092 (-0.33, 0.15)

## forecast based on priors only
mod_unfitted <- mod_pois(injuries ~ age * sex + ethnicity + year,
                         data = injuries,
                         exposure = popn)
mod_unfitted |>
  forecast(labels = 2019:2024)
#> ℹ Model not fitted, so values drawn straight from prior distribution.
#> # A tibble: 288 × 9
#>    age   sex    ethnicity  year injuries  popn .observed               .fitted
#>    <fct> <chr>  <chr>     <int>    <int> <int>     <dbl>          <rdbl<1000>>
#>  1 0-4   Female Maori      2019       NA    NA        NA  0.67 (1.3e-05, 8011)
#>  2 0-4   Female Maori      2020       NA    NA        NA    0.68 (2e-05, 9945)
#>  3 0-4   Female Maori      2021       NA    NA        NA 0.73 (9.4e-06, 15674)
#>  4 0-4   Female Maori      2022       NA    NA        NA 0.61 (5.5e-06, 12882)
#>  5 0-4   Female Maori      2023       NA    NA        NA 0.63 (4.2e-06, 38775)
#>  6 0-4   Female Maori      2024       NA    NA        NA 0.64 (3.1e-06, 53122)
#>  7 0-4   Female Non Maori  2019       NA    NA        NA  0.6 (1.6e-05, 12380)
#>  8 0-4   Female Non Maori  2020       NA    NA        NA 0.69 (9.4e-06, 12572)
#>  9 0-4   Female Non Maori  2021       NA    NA        NA 0.68 (4.6e-06, 15645)
#> 10 0-4   Female Non Maori  2022       NA    NA        NA  0.7 (1.1e-05, 22968)
#> # ℹ 278 more rows
#> # ℹ 1 more variable: .expected <rdbl<1000>>