Create components needed by a ssvd object

Use the Singular Value Decomposition (SVD) to construct a parsimonious representation of a set of rates, probabilities, means, or other values.

Usage

ssvd_comp(x, transform = c("log", "logit", "none"), n_comp = 10)

Arguments

x

A matrix with value such as rates, probabilities, or means.

transform

"log", "logit", or "none". Defaults to "log".

n_comp

Number of components.

Value

A named list with two elements:

matrix, a numeric matrix

offset, a numeric vector

Details

`ssvd_comp() typically proceeds as follows:

transform values in matrix x (eg take logs)

carry out a SVD on the transformed version of x

centre and scale the results from the SVD to produce matrix matrix and vector offset.

If

\(X\) is a the matrix of transformed values of x

\(F\) is the matrix called matrix,

\(g\) is the vector called offset, and

\(\alpha\) is a vector of standard normal variates,

and

$$\beta = F \alpha + g$$

then \(\beta\) should look like a randomly-selected column from \(X\).

Matrix x typically has age along the rows, and some combination of classification variables, such as country and time, along the columns. One exception is when the SVD is used to capture the relationship between female and male rates, in which case rows are formed by interacting sex and age. See below for an example.

When scale is "log" or "logit", ssvd_comp() converts any 0s in x to values just above 0 before applying the log or logit function. When scale is "logit", ssvd_comp() also converts any 1s to values just below 1`.

Examples

x <- matrix(rgamma(n = 150, shape = 1),
            nrow = 10,
            ncol = 15)
x
#>            [,1]       [,2]      [,3]      [,4]       [,5]       [,6]       [,7]
#>  [1,] 0.3461946 1.03866812 0.8579223 2.4071455 0.22348067 1.29610902 2.26126133
#>  [2,] 0.3808581 4.33830316 1.5465257 3.8493147 0.03111293 0.24186057 1.62900859
#>  [3,] 0.1322578 0.54018009 0.7448412 1.0412970 0.10103279 0.83996486 1.61608163
#>  [4,] 3.3337419 3.12327092 1.1877186 0.7645732 0.13476688 0.58809071 0.22140288
#>  [5,] 1.9723722 0.95796616 0.3339474 3.9915602 0.09195434 0.29414449 1.16626890
#>  [6,] 0.4721629 0.93864619 1.2104465 0.6163676 0.75614498 0.23744885 3.81129852
#>  [7,] 2.2137929 1.08033117 0.7900078 0.6284162 0.33953181 0.28484718 0.09012904
#>  [8,] 1.4950270 0.24336074 0.1557304 0.3548268 2.64119344 0.31428904 1.50147280
#>  [9,] 0.3364508 0.71140853 1.6842143 2.4260707 0.28416797 0.04176534 0.53004726
#> [10,] 0.5694015 0.03882585 1.0423669 1.7074622 0.30445829 1.54406522 0.02112217
#>            [,8]        [,9]      [,10]      [,11]      [,12]       [,13]
#>  [1,] 0.3822741 0.358186804 1.03017219 0.13138214 1.59065031 1.783046793
#>  [2,] 0.2311672 1.708678861 0.71605141 1.06306892 1.05826921 0.790463654
#>  [3,] 1.6914218 0.006869335 1.41342362 0.66824690 0.84118473 0.825461866
#>  [4,] 0.1318079 1.761283438 0.06667042 3.78653029 1.81067276 0.160614300
#>  [5,] 3.1174828 4.853258682 1.19781461 1.11596393 0.98859076 0.411382575
#>  [6,] 0.6635328 1.229702735 0.18267094 2.13337871 0.64599525 0.007829091
#>  [7,] 0.7096522 0.184874734 0.08703218 1.28287837 0.02297977 0.702652982
#>  [8,] 1.1517015 0.206969435 0.09997737 0.55639790 4.28366827 2.336699324
#>  [9,] 1.9698156 1.460254436 1.46346324 0.07726084 2.03069035 0.332754370
#> [10,] 0.2820437 0.540422716 0.21811531 0.02305563 0.89419669 0.423643541
#>            [,14]     [,15]
#>  [1,] 2.27439929 0.8925609
#>  [2,] 0.07845103 0.2282912
#>  [3,] 0.27417139 0.4914270
#>  [4,] 1.73025638 0.3643840
#>  [5,] 2.31180756 0.7661803
#>  [6,] 0.72931908 0.1156978
#>  [7,] 0.93378084 0.2319514
#>  [8,] 0.25141280 1.3064909
#>  [9,] 0.73427656 1.1965033
#> [10,] 0.77107322 0.2114691
ssvd_comp(x)
#> $matrix
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                           
#>  [1,] 0.1673454  0.41030164  0.4292109  0.20143535  0.13715794  0.07662589
#>  [2,] 0.4618001 -0.45058292  0.8229403  0.11607090  0.46592816  0.39045169
#>  [3,] 0.4186056  0.77078809  0.8040155 -0.82586076  0.05703270 -0.17911507
#>  [4,] 0.4370270 -0.82028442 -0.4093244 -0.03208771  0.20041534  0.46942760
#>  [5,] 0.1611690 -0.44435467  0.2843709  0.47744379  0.10312800 -0.05561233
#>  [6,] 0.4714580 -0.92069392  0.1058578 -0.30635077 -0.83918158 -0.34127873
#>  [7,] 0.5671986 -0.07782588 -0.6819546 -0.56964097  0.50044600 -0.27980234
#>  [8,] 0.1043730  0.47179738 -0.1727047 -0.41559238 -0.59253095  0.77696912
#>  [9,] 0.3603144  0.14867128  0.4974068  0.67875732 -0.34495230 -0.11228490
#> [10,] 0.7158273  0.86929276 -0.6700956  0.64514843 -0.09538405 -0.04942423
#>                                                      
#>  [1,] -0.26231989 -0.26996470 -0.25614383  0.37597340
#>  [2,] -0.07851500  0.38400166  0.17692192  0.06771365
#>  [3,] -0.09545652 -0.20314908  0.02483615 -0.20148864
#>  [4,] -0.25250096 -0.25781023 -0.23008879 -0.20476221
#>  [5,]  0.43979582 -0.50415683  0.27918124  0.02263371
#>  [6,] -0.21929427  0.03259973  0.05281229  0.11930027
#>  [7,]  0.48642421  0.09336759 -0.07930730  0.14373294
#>  [8,]  0.35499629 -0.02044537  0.06251044  0.09065075
#>  [9,]  0.47320865  0.15956818 -0.31084305 -0.13296436
#> [10,] -0.30801261  0.03229974  0.18424290 -0.03258780
#> 
#> $offset
#>  [1] -0.197231416 -0.501209667 -0.784071580 -0.421968022  0.009749737
#>  [6] -0.692738775 -0.963333035 -0.454625164 -0.449037316 -1.187039031
#> 
ssvd_comp(x, transform = "none")
#> $matrix
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                            
#>  [1,] 0.4327321  0.34366846  0.27062062  0.34150655  0.01171887 -0.47963892
#>  [2,] 0.5562370 -0.41403530 -0.02193286  0.75807996  0.37209155  0.27781250
#>  [3,] 0.2704592  0.24582460  0.20855937  0.18433888 -0.16933802 -0.10495549
#>  [4,] 0.5302481 -0.30255736 -1.05719132 -0.25130029  0.10215106 -0.08120471
#>  [5,] 0.6986508 -0.53081339  0.59531080 -0.62676374 -0.19110664  0.08976653
#>  [6,] 0.3684973  0.12181192 -0.23197540  0.35473516 -0.78525077  0.09627665
#>  [7,] 0.2417867 -0.08533488 -0.38701627 -0.18009102  0.05670324 -0.21964220
#>  [8,] 0.3832416  1.20785763 -0.18530928 -0.25334017  0.15413364  0.25306357
#>  [9,] 0.4168738  0.10756791  0.46948497 -0.07801618  0.20003456  0.11783604
#> [10,] 0.2186941  0.06577901  0.17241450 -0.02620374  0.21171734 -0.30509938
#>                                                        
#>  [1,]  0.190793238 -0.02200785 -0.153996299 -0.04142857
#>  [2,]  0.087893305 -0.06327824  0.053098220  0.01051819
#>  [3,] -0.330896254 -0.16865020 -0.004448699  0.16154519
#>  [4,] -0.060675253  0.10053406 -0.111880720  0.05689472
#>  [5,]  0.160834863 -0.06206493  0.005255647  0.03272854
#>  [6,] -0.001499619  0.10860803  0.069035031 -0.05221435
#>  [7,] -0.108009890 -0.23760226  0.151285601 -0.14126579
#>  [8,]  0.121512350 -0.03655492  0.035171650  0.01374517
#>  [9,] -0.342439304  0.15108476 -0.078922141 -0.12093375
#> [10,]  0.003205145  0.22333155  0.234862671  0.06353519
#> 
#> $offset
#>  [1] 1.1248969 1.1927617 0.7485241 1.2777190 1.5713797 0.9167095 0.6388572
#>  [8] 1.1266145 1.0186096 0.5727815
#> 
ssvd_comp(x, n_comp = 2)
#> $matrix
#> 10 x 2 sparse Matrix of class "dgCMatrix"
#>                            
#>  [1,] 0.1673454  0.41030164
#>  [2,] 0.4618001 -0.45058292
#>  [3,] 0.4186056  0.77078809
#>  [4,] 0.4370270 -0.82028442
#>  [5,] 0.1611690 -0.44435467
#>  [6,] 0.4714580 -0.92069392
#>  [7,] 0.5671986 -0.07782588
#>  [8,] 0.1043730  0.47179738
#>  [9,] 0.3603144  0.14867128
#> [10,] 0.7158273  0.86929276
#> 
#> $offset
#>  [1] -0.3273475 -0.5953393 -0.7691142 -0.4867607 -0.1519088 -0.5181469
#>  [7] -0.8237598 -0.2461755 -0.5618450 -1.2278814
#>